pyproject.toml
[tool.poetry]
name = "api-pipeline-with-agent"
version = "0.1.0"
description = ""
authors = ["Rafael Panegassi <rm.panegassi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.0"
telethon = "^1.40.0"
python-dotenv = "^1.1.0"
loguru = "^0.7.3"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
